@@ -1,6 +1,6 @@
 {
   "name": "vis",
-  "version": "4.14.0",
+  "version": "4.16.1",
   "description": "A dynamic, browser-based visualization library.",
   "homepage": "http://visjs.org/",
   "license": "(Apache-2.0 OR MIT)",
@@ -30,26 +30,28 @@
   },
   "dependencies": {},
   "devDependencies": {
+    "async": "^2.0.0-rc.2",
+    "babel-core": "^6.6.5",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.6.0",
+    "babelify": "^7.2.0",
+    "clean-css": "^3.4.10",
     "emitter-component": "^1.1.1",
+    "gulp": "^3.9.1",
+    "gulp-concat": "^2.6.0",
+    "gulp-minify-css": "^1.2.4",
+    "gulp-rename": "^1.2.2",
+    "gulp-util": "^3.0.7",
     "hammerjs": "^2.0.6",
     "keycharm": "^0.2.0",
-    "moment": "^2.10.2",
-    "propagating-hammerjs": "^1.4.5",
+    "merge-stream": "^1.0.0",
+    "mocha": "^2.4.5",
+    "moment": "^2.12.0",
+    "propagating-hammerjs": "^1.4.6",
+    "rimraf": "^2.5.2",
+    "uglify-js": "^2.6.2",
     "uuid": "^2.0.1",
-    "babel": "^5.1.11",
-    "babel-loader": "^5.0.0",
-    "babelify": "^6.0.2",
-    "clean-css": "^3.2.1",
-    "gulp": "^3.8.11",
-    "gulp-concat": "^2.5.2",
-    "gulp-minify-css": "^1.0.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-util": "^3.0.4",
-    "merge-stream": "^0.1.7",
-    "mocha": "^2.2.4",
-    "rimraf": "^2.3.2",
-    "uglify-js": "^2.4.20",
-    "webpack": "^1.8.5",
-    "yargs": "^3.7.2"
+    "webpack": "^1.12.14",
+    "yargs": "^4.2.0"
   }
 }
