@@ -1,5 +1,6 @@
-let Node = require("./components/Node");
-let Edge = require("./components/Edge");
+import Node from './components/Node';
+import Edge from './components/Edge';
+
 let util = require('../../util');
 
 class SelectionHandler {
