@@ -2237,9 +2237,9 @@ Graph3d.prototype._showTooltip = function (dataPoint) {
   }
   else {
     content.innerHTML = '<table>' +
-      '<tr><td>x:</td><td>' + dataPoint.point.x + '</td></tr>' +
-      '<tr><td>y:</td><td>' + dataPoint.point.y + '</td></tr>' +
-      '<tr><td>z:</td><td>' + dataPoint.point.z + '</td></tr>' +
+      '<tr><td>' + this.xLabel + ':</td><td>' + dataPoint.point.x + '</td></tr>' +
+      '<tr><td>' + this.yLabel + ':</td><td>' + dataPoint.point.y + '</td></tr>' +
+      '<tr><td>' + this.zLabel + ':</td><td>' + dataPoint.point.z + '</td></tr>' +
       '</table>';
   }
 
