@@ -2,16 +2,15 @@
 require('./shapes');
 
 let Emitter = require('emitter-component');
-let Hammer = require('../module/hammer');
 let util = require('../util');
 let DataSet = require('../DataSet');
 let DataView = require('../DataView');
 let dotparser = require('./dotparser');
 let gephiParser = require('./gephiParser');
-let Images = require('./Images');
 let Activator = require('../shared/Activator');
 let locales = require('./locales');
 
+import  Images              from './Images';
 import  Groups              from './modules/Groups';
 import  NodesHandler        from './modules/NodesHandler';
 import  EdgesHandler        from './modules/EdgesHandler';
