@@ -17,6 +17,7 @@ class Canvas {
     this.resizeTimer = undefined;
     this.resizeFunction = this._onResize.bind(this);
     this.cameraState = {};
+    this.initialized = false;
 
     this.options = {};
     this.defaultOptions = {
@@ -88,10 +89,15 @@ class Canvas {
    * @private
    */
   _getCameraState(pixelRatio = this.pixelRatio) {
-    this.cameraState.previousWidth = this.frame.canvas.width / pixelRatio;
-    this.cameraState.previousHeight = this.frame.canvas.height / pixelRatio;
-    this.cameraState.scale = this.body.view.scale;
-    this.cameraState.position = this.DOMtoCanvas({x: 0.5 * this.frame.canvas.width / pixelRatio, y: 0.5 * this.frame.canvas.height / pixelRatio});
+    if (this.initialized === true) {
+      this.cameraState.previousWidth = this.frame.canvas.width / pixelRatio;
+      this.cameraState.previousHeight = this.frame.canvas.height / pixelRatio;
+      this.cameraState.scale = this.body.view.scale;
+      this.cameraState.position = this.DOMtoCanvas({
+        x: 0.5 * this.frame.canvas.width / pixelRatio,
+        y: 0.5 * this.frame.canvas.height / pixelRatio
+      });
+    }
   }
 
   /**
@@ -216,7 +222,7 @@ class Canvas {
     this.hammer = new Hammer(this.frame.canvas);
     this.hammer.get('pinch').set({enable: true});
     // enable to get better response, todo: test on mobile.
-    this.hammer.get('pan').set({threshold:5, direction:30}); // 30 is ALL_DIRECTIONS in hammer.
+    this.hammer.get('pan').set({threshold:5, direction: Hammer.DIRECTION_ALL});
 
     hammerUtil.onTouch(this.hammer, (event) => {this.body.eventListeners.onTouch(event)});
     this.hammer.on('tap',       (event) => {this.body.eventListeners.onTap(event)});
@@ -311,6 +317,9 @@ class Canvas {
       this._setCameraState();
     }
 
+
+    // set initialized so the get and set camera will work from now on.
+    this.initialized = true;
     return emitEvent;
   };
 
