@@ -23,6 +23,7 @@ class LayoutEngine {
         treeSpacing: 200,
         blockShifting: true,
         edgeMinimization: true,
+        parentCentralization: true,
         direction: 'UD',   // UD, DU, LR, RL
         sortMethod: 'hubsize' // hubsize, directed
       }
@@ -294,8 +295,8 @@ class LayoutEngine {
       let undefinedLevel = false;
       this.hierarchicalLevels = {};
       this.lastNodeOnLevel = {};
-      this.hierarchicalParents = {};
-      this.hierarchicalChildren = {};
+      this.hierarchicalChildrenReference = {};
+      this.hierarchicalParentReference = {};
       this.hierarchicalTrees = {};
       this.treeIndex = -1;
 
@@ -379,9 +380,7 @@ class LayoutEngine {
       let treeSizes = getTreeSizes();
       for (let i = 0; i < treeSizes.length - 1; i++) {
         let diff = treeSizes[i].max - treeSizes[i+1].min;
-        if (diff !== this.options.hierarchical.treeSpacing) {
-          shiftTree(i + 1, diff - this.options.hierarchical.treeSpacing);
-        }
+        shiftTree(i + 1, diff + this.options.hierarchical.treeSpacing);
       }
     };
 
@@ -390,7 +389,9 @@ class LayoutEngine {
       for (let nodeId in this.hierarchicalTrees) {
         if (this.hierarchicalTrees.hasOwnProperty(nodeId)) {
           if (this.hierarchicalTrees[nodeId] === index) {
-            this._setPositionForHierarchy(this.body.nodes[nodeId], offset, undefined, true);
+            let node = this.body.nodes[nodeId];
+            let pos = this._getPositionForHierarchy(node);
+            this._setPositionForHierarchy(node, pos + offset, undefined, true);
           }
         }
       }
@@ -415,7 +416,7 @@ class LayoutEngine {
     // get the width of all trees
     let getTreeSizes = () => {
       let treeWidths = [];
-      for (let i = 0; i < this.treeIndex; i++) {
+      for (let i = 0; i <= this.treeIndex; i++) {
         treeWidths.push(getTreeSize(i));
       }
       return treeWidths;
@@ -425,8 +426,8 @@ class LayoutEngine {
     // get a map of all nodes in this branch
     let getBranchNodes = (source, map) => {
       map[source.id] = true;
-      if (this.hierarchicalParents[source.id]) {
-        let children = this.hierarchicalParents[source.id].children;
+      if (this.hierarchicalChildrenReference[source.id]) {
+        let children = this.hierarchicalChildrenReference[source.id];
         if (children.length > 0) {
           for (let i = 0; i < children.length; i++) {
             getBranchNodes(this.body.nodes[children[i]], map);
@@ -467,8 +468,8 @@ class LayoutEngine {
     // get the maximum level of a branch.
     let getMaxLevel = (nodeId) => {
       let level = this.hierarchicalLevels[nodeId];
-      if (this.hierarchicalParents[nodeId]) {
-        let children = this.hierarchicalParents[nodeId].children;
+      if (this.hierarchicalChildrenReference[nodeId]) {
+        let children = this.hierarchicalChildrenReference[nodeId];
         if (children.length > 0) {
           for (let i = 0; i < children.length; i++) {
             level = Math.max(level,getMaxLevel(children[i]));
@@ -487,13 +488,12 @@ class LayoutEngine {
 
     // check if two nodes have the same parent(s)
     let hasSameParent = (node1, node2) => {
-      let parents1 = this.hierarchicalChildren[node1.id];
-      let parents2 = this.hierarchicalChildren[node2.id];
+      let parents1 = this.hierarchicalParentReference[node1.id];
+      let parents2 = this.hierarchicalParentReference[node2.id];
       if (parents1 === undefined || parents2 === undefined) {
         return false;
       }
-      parents1 = parents1.parents;
-      parents2 = parents2.parents;
+
       for (let i = 0; i < parents1.length; i++) {
         for (let j = 0; j < parents2.length; j++) {
           if (parents1[i] == parents2[j]) {
@@ -676,7 +676,7 @@ class LayoutEngine {
           }
 
           if (newPosition !== nodePosition) {
-            //console.log("moving Node:",diff, minSpace, maxSpace)
+            //console.log("moving Node:",diff, minSpace, maxSpace);
             this._setPositionForHierarchy(node, newPosition, undefined, true);
             //this.body.emitter.emit("_redraw");
             stillShifting = true;
@@ -710,7 +710,7 @@ class LayoutEngine {
       }
     };
 
-    //// method to remove whitespace between branches. Because we do bottom up, we can center the parents.
+    // method to remove whitespace between branches. Because we do bottom up, we can center the parents.
     let shiftBranchesCloserBottomUp = (iterations) => {
       let levels = Object.keys(this.distributionOrdering);
       levels = levels.reverse();
@@ -732,6 +732,19 @@ class LayoutEngine {
       }
     };
 
+    // center all parents
+    let centerAllParentsBottomUp = () => {
+      let levels = Object.keys(this.distributionOrdering);
+      levels = levels.reverse();
+      for (let i = 0; i < levels.length; i++) {
+        let level = levels[i];
+        let levelNodes = this.distributionOrdering[level];
+        for (let j = 0; j < levelNodes.length; j++) {
+          this._centerParent(levelNodes[j]);
+        }
+      }
+    };
+
     // the actual work is done here.
     if (this.options.hierarchical.blockShifting === true) {
       shiftBranchesCloserBottomUp(5);
@@ -741,6 +754,10 @@ class LayoutEngine {
     // minimize edge length
     if (this.options.hierarchical.edgeMinimization === true) {
       minimizeEdgeLengthBottomUp(20);
+    }
+
+    if (this.options.hierarchical.parentCentralization === true) {
+      centerAllParentsBottomUp()
     }
 
     shiftTrees();
@@ -794,16 +811,16 @@ class LayoutEngine {
    * @private
    */
   _centerParent(node) {
-    if (this.hierarchicalChildren[node.id]) {
-      let parents = this.hierarchicalChildren[node.id].parents;
+    if (this.hierarchicalParentReference[node.id]) {
+      let parents = this.hierarchicalParentReference[node.id];
       for (var i = 0; i < parents.length; i++) {
         let parentId = parents[i];
         let parentNode = this.body.nodes[parentId];
-        if (this.hierarchicalParents[parentId]) {
+        if (this.hierarchicalChildrenReference[parentId]) {
           // get the range of the children
           let minPos = 1e9;
           let maxPos = -1e9;
-          let children = this.hierarchicalParents[parentId].children;
+          let children = this.hierarchicalChildrenReference[parentId];
           if (children.length > 0) {
             for (let i = 0; i < children.length; i++) {
               let childNode = this.body.nodes[children[i]];
@@ -841,17 +858,105 @@ class LayoutEngine {
         let nodeArray = Object.keys(distribution[level]);
         nodeArray = this._indexArrayToNodes(nodeArray);
         this._sortNodeArray(nodeArray);
+        let handledNodeCount = 0;
 
         for (let i = 0; i < nodeArray.length; i++) {
           let node = nodeArray[i];
           if (this.positionedNodes[node.id] === undefined) {
-            this._setPositionForHierarchy(node, this.options.hierarchical.nodeSpacing * i, level);
-            this.positionedNodes[node.id] = true;
-            this._placeBranchNodes(node.id, level);
+            let pos = this.options.hierarchical.nodeSpacing * handledNodeCount;
+            // we get the X or Y values we need and store them in pos and previousPos. The get and set make sure we get X or Y
+            if (handledNodeCount > 0) {pos = this._getPositionForHierarchy(nodeArray[i-1]) + this.options.hierarchical.nodeSpacing;}
+            this._setPositionForHierarchy(node, pos, level);
+            this._validataPositionAndContinue(node, level, pos);
+
+            handledNodeCount++;
           }
         }
       }
     }
+  }
+
+
+  /**
+   * This is a recursively called function to enumerate the branches from the largest hubs and place the nodes
+   * on a X position that ensures there will be no overlap.
+   *
+   * @param parentId
+   * @param parentLevel
+   * @private
+   */
+  _placeBranchNodes(parentId, parentLevel) {
+    // if this is not a parent, cancel the placing. This can happen with multiple parents to one child.
+    if (this.hierarchicalChildrenReference[parentId] === undefined) {
+      return;
+    }
+
+    // get a list of childNodes
+    let childNodes = [];
+    for (let i = 0; i < this.hierarchicalChildrenReference[parentId].length; i++) {
+      childNodes.push(this.body.nodes[this.hierarchicalChildrenReference[parentId][i]]);
+    }
+
+    // use the positions to order the nodes.
+    this._sortNodeArray(childNodes);
+
+    // position the childNodes
+    for (let i = 0; i < childNodes.length; i++) {
+      let childNode = childNodes[i];
+      let childNodeLevel = this.hierarchicalLevels[childNode.id];
+      // check if the child node is below the parent node and if it has already been positioned.
+      if (childNodeLevel > parentLevel && this.positionedNodes[childNode.id] === undefined) {
+        // get the amount of space required for this node. If parent the width is based on the amount of children.
+        let pos;
+
+        // we get the X or Y values we need and store them in pos and previousPos. The get and set make sure we get X or Y
+        if (i === 0) {pos = this._getPositionForHierarchy(this.body.nodes[parentId]);}
+        else         {pos = this._getPositionForHierarchy(childNodes[i-1]) + this.options.hierarchical.nodeSpacing;}
+        this._setPositionForHierarchy(childNode, pos, childNodeLevel);
+        this._validataPositionAndContinue(childNode, childNodeLevel, pos);
+      }
+      else {
+        return;
+      }
+    }
+
+    // center the parent nodes.
+    let minPos = 1e9;
+    let maxPos = -1e9;
+    for (let i = 0; i < childNodes.length; i++) {
+      let childNodeId = childNodes[i].id;
+      minPos = Math.min(minPos, this._getPositionForHierarchy(this.body.nodes[childNodeId]));
+      maxPos = Math.max(maxPos, this._getPositionForHierarchy(this.body.nodes[childNodeId]));
+    }
+    this._setPositionForHierarchy(this.body.nodes[parentId], 0.5 * (minPos + maxPos), parentLevel);
+  }
+
+
+  /**
+   * This method checks for overlap and if required shifts the branch. It also keeps records of positioned nodes.
+   * Finally it will call _placeBranchNodes to place the branch nodes.
+   * @param node
+   * @param level
+   * @param pos
+   * @private
+   */
+  _validataPositionAndContinue(node, level, pos) {
+    // if overlap has been detected, we shift the branch
+    if (this.lastNodeOnLevel[level] !== undefined) {
+      let previousPos = this._getPositionForHierarchy(this.body.nodes[this.lastNodeOnLevel[level]]);
+      if (pos - previousPos < this.options.hierarchical.nodeSpacing) {
+        let diff = (previousPos + this.options.hierarchical.nodeSpacing) - pos;
+        let sharedParent = this._findCommonParent(this.lastNodeOnLevel[level], node.id);
+        this._shiftBlock(sharedParent.withChild, diff);
+      }
+    }
+
+    // store change in position.
+    this.lastNodeOnLevel[level] = node.id;
+
+    this.positionedNodes[node.id] = true;
+
+    this._placeBranchNodes(node.id, level);
   }
 
   /**
@@ -1047,14 +1152,14 @@ class LayoutEngine {
       if (this.hierarchicalLevels[childNode.id] > this.hierarchicalLevels[parentNode.id]) {
         let parentNodeId = parentNode.id;
         let childNodeId = childNode.id;
-        if (this.hierarchicalParents[parentNodeId] === undefined) {
-          this.hierarchicalParents[parentNodeId] = {children: [], amount: 0};
+        if (this.hierarchicalChildrenReference[parentNodeId] === undefined) {
+          this.hierarchicalChildrenReference[parentNodeId] = [];
         }
-        this.hierarchicalParents[parentNodeId].children.push(childNodeId);
-        if (this.hierarchicalChildren[childNodeId] === undefined) {
-          this.hierarchicalChildren[childNodeId] = {parents: [], amount: 0};
+        this.hierarchicalChildrenReference[parentNodeId].push(childNodeId);
+        if (this.hierarchicalParentReference[childNodeId] === undefined) {
+          this.hierarchicalParentReference[childNodeId] = [];
         }
-        this.hierarchicalChildren[childNodeId].parents.push(parentNodeId);
+        this.hierarchicalParentReference[childNodeId].push(parentNodeId);
       }
     };
 
@@ -1070,8 +1175,16 @@ class LayoutEngine {
    */
   _crawlNetwork(callback = function() {}, startingNodeId) {
     let progress = {};
-    let crawler = (node) => {
+    let treeIndex = 0;
+
+    let crawler = (node, tree) => {
       if (progress[node.id] === undefined) {
+
+        if (this.hierarchicalTrees[node.id] === undefined) {
+          this.hierarchicalTrees[node.id] = tree;
+          this.treeIndex = Math.max(tree, this.treeIndex);
+        }
+
         progress[node.id] = true;
         let childNode;
         for (let i = 0; i < node.edges.length; i++) {
@@ -1085,7 +1198,7 @@ class LayoutEngine {
 
             if (node.id !== childNode.id) {
               callback(node, childNode, node.edges[i]);
-              crawler(childNode);
+              crawler(childNode, tree);
             }
           }
         }
@@ -1097,7 +1210,10 @@ class LayoutEngine {
     if (startingNodeId === undefined) {
       for (let i = 0; i < this.body.nodeIndices.length; i++) {
         let node = this.body.nodes[this.body.nodeIndices[i]];
-        crawler(node);
+        if (progress[node.id] === undefined) {
+          crawler(node, treeIndex);
+          treeIndex += 1;
+        }
       }
     }
     else {
@@ -1108,77 +1224,6 @@ class LayoutEngine {
       }
       crawler(node);
     }
-  }
-
-
-  /**
-   * This is a recursively called function to enumerate the branches from the largest hubs and place the nodes
-   * on a X position that ensures there will be no overlap.
-   *
-   * @param parentId
-   * @param parentLevel
-   * @private
-   */
-  _placeBranchNodes(parentId, parentLevel) {
-    // if this is not a parent, cancel the placing. This can happen with multiple parents to one child.
-    if (this.hierarchicalParents[parentId] === undefined) {
-      return;
-    }
-
-    // get a list of childNodes
-    let childNodes = [];
-    for (let i = 0; i < this.hierarchicalParents[parentId].children.length; i++) {
-      childNodes.push(this.body.nodes[this.hierarchicalParents[parentId].children[i]]);
-    }
-
-    // use the positions to order the nodes.
-    this._sortNodeArray(childNodes);
-
-    // position the childNodes
-    for (let i = 0; i < childNodes.length; i++) {
-      let childNode = childNodes[i];
-      let childNodeLevel = this.hierarchicalLevels[childNode.id];
-      // check if the child node is below the parent node and if it has already been positioned.
-      if (childNodeLevel > parentLevel && this.positionedNodes[childNode.id] === undefined) {
-        // get the amount of space required for this node. If parent the width is based on the amount of children.
-        let pos;
-
-        // we get the X or Y values we need and store them in pos and previousPos. The get and set make sure we get X or Y
-        if (i === 0) {pos = this._getPositionForHierarchy(this.body.nodes[parentId]);}
-        else         {pos = this._getPositionForHierarchy(childNodes[i-1]) + this.options.hierarchical.nodeSpacing;}
-        this._setPositionForHierarchy(childNode, pos, childNodeLevel);
-
-        // if overlap has been detected, we shift the branch
-        if (this.lastNodeOnLevel[childNodeLevel] !== undefined) {
-          let previousPos = this._getPositionForHierarchy(this.body.nodes[this.lastNodeOnLevel[childNodeLevel]]);
-          if (pos - previousPos < this.options.hierarchical.nodeSpacing) {
-            let diff = (previousPos + this.options.hierarchical.nodeSpacing) - pos;
-            let sharedParent = this._findCommonParent(this.lastNodeOnLevel[childNodeLevel], childNode.id);
-            this._shiftBlock(sharedParent.withChild, diff);
-          }
-        }
-
-        // store change in position.
-        this.lastNodeOnLevel[childNodeLevel] = childNode.id;
-
-        this.positionedNodes[childNode.id] = true;
-
-        this._placeBranchNodes(childNode.id, childNodeLevel);
-      }
-      else {
-        return;
-      }
-    }
-
-    // center the parent nodes.
-    let minPos = 1e9;
-    let maxPos = -1e9;
-    for (let i = 0; i < childNodes.length; i++) {
-      let childNodeId = childNodes[i].id;
-      minPos = Math.min(minPos, this._getPositionForHierarchy(this.body.nodes[childNodeId]));
-      maxPos = Math.max(maxPos, this._getPositionForHierarchy(this.body.nodes[childNodeId]));
-    }
-    this._setPositionForHierarchy(this.body.nodes[parentId], 0.5 * (minPos + maxPos), parentLevel);
   }
 
 
@@ -1195,9 +1240,9 @@ class LayoutEngine {
     else {
       this.body.nodes[parentId].y += diff;
     }
-    if (this.hierarchicalParents[parentId] !== undefined) {
-      for (let i = 0; i < this.hierarchicalParents[parentId].children.length; i++) {
-        this._shiftBlock(this.hierarchicalParents[parentId].children[i], diff);
+    if (this.hierarchicalChildrenReference[parentId] !== undefined) {
+      for (let i = 0; i < this.hierarchicalChildrenReference[parentId].length; i++) {
+        this._shiftBlock(this.hierarchicalChildrenReference[parentId][i], diff);
       }
     }
   }
@@ -1213,18 +1258,18 @@ class LayoutEngine {
   _findCommonParent(childA,childB) {
     let parents = {};
     let iterateParents = (parents,child) => {
-      if (this.hierarchicalChildren[child] !== undefined) {
-        for (let i = 0; i < this.hierarchicalChildren[child].parents.length; i++) {
-          let parent = this.hierarchicalChildren[child].parents[i];
+      if (this.hierarchicalParentReference[child] !== undefined) {
+        for (let i = 0; i < this.hierarchicalParentReference[child].length; i++) {
+          let parent = this.hierarchicalParentReference[child][i];
           parents[parent] = true;
           iterateParents(parents, parent)
         }
       }
     };
     let findParent = (parents, child) => {
-      if (this.hierarchicalChildren[child] !== undefined) {
-        for (let i = 0; i < this.hierarchicalChildren[child].parents.length; i++) {
-          let parent = this.hierarchicalChildren[child].parents[i];
+      if (this.hierarchicalParentReference[child] !== undefined) {
+        for (let i = 0; i < this.hierarchicalParentReference[child].length; i++) {
+          let parent = this.hierarchicalParentReference[child][i];
           if (parents[parent] !== undefined) {
             return {foundParent:parent, withChild:child};
           }
@@ -1249,6 +1294,7 @@ class LayoutEngine {
    * @private
    */
   _setPositionForHierarchy(node, position, level, doNotUpdate = false) {
+    //console.log('_setPositionForHierarchy',node.id, position)
     if (doNotUpdate !== true) {
       if (this.distributionOrdering[level] === undefined) {
         this.distributionOrdering[level] = [];
@@ -1260,28 +1306,6 @@ class LayoutEngine {
         this.distributionIndex[node.id] = this.distributionOrdering[level].length - 1;
       }
       this.distributionOrderingPresence[level][node.id] = true;
-
-      if (this.hierarchicalTrees[node.id] === undefined) {
-        if (this.hierarchicalChildren[node.id] !== undefined) {
-          let tree = 1;
-          // get the lowest tree denominator.
-          for (let i = 0; i < this.hierarchicalChildren[node.id].parents.length; i++) {
-            let parentId = this.hierarchicalChildren[node.id].parents[i];
-            if (this.hierarchicalTrees[parentId] !== undefined) {
-              //tree = Math.min(tree,this.hierarchicalTrees[parentId]);
-              tree = this.hierarchicalTrees[parentId];
-            }
-          }
-          //for (let i = 0; i < this.hierarchicalChildren.parents.length; i++) {
-          //  let parentId = this.hierarchicalChildren.parents[i];
-          //  this.hierarchicalTrees[parentId] = tree;
-          //}
-          this.hierarchicalTrees[node.id] = tree;
-        }
-        else {
-          this.hierarchicalTrees[node.id] = ++this.treeIndex;
-        }
-      }
     }
 
     if (this.options.hierarchical.direction === 'UD' || this.options.hierarchical.direction === 'DU') {
