@@ -263,7 +263,7 @@ class ManipulationSystem {
       this.inMode = 'editNode';
       if (typeof this.options.editNode === 'function') {
         if (node.isCluster !== true) {
-          let data = util.deepExtend({}, node.options, true);
+          let data = util.deepExtend({}, node.options, false);
           data.x = node.x;
           data.y = node.y;
 
@@ -689,7 +689,12 @@ class ManipulationSystem {
   }
 
   _createDeleteButton(locale) {
-    let button = this._createButton('delete', 'vis-button vis-delete', locale['del'] || this.options.locales['en']['del']);
+    if (this.options.rtl) {
+      var deleteBtnClass = 'vis-button vis-delete-rtl';
+    } else {
+      var deleteBtnClass = 'vis-button vis-delete';
+    }
+    let button = this._createButton('delete', deleteBtnClass, locale['del'] || this.options.locales['en']['del']);
     this.manipulationDiv.appendChild(button);
     this._bindHammerToDiv(button, this.deleteSelected.bind(this));
   }
@@ -848,6 +853,11 @@ class ManipulationSystem {
       edge.edgeType.to = to;
     }
 
+    // we use the selection to find the node that is being dragged. We explicitly select it here.
+    if (this.selectedControlNode !== undefined) {
+      this.selectionHandler.selectObject(this.selectedControlNode)
+    }
+
     this.body.emitter.emit('_redraw');
   }
 
@@ -860,7 +870,6 @@ class ManipulationSystem {
     this.body.emitter.emit('disablePhysics');
     let pointer = this.body.functions.getPointer(event.center);
     let pos = this.canvas.DOMtoCanvas(pointer);
-
     if (this.selectedControlNode !== undefined) {
       this.selectedControlNode.x = pos.x;
       this.selectedControlNode.y = pos.y;
@@ -884,12 +893,13 @@ class ManipulationSystem {
     let pointer = this.body.functions.getPointer(event.center);
     let pointerObj = this.selectionHandler._pointerToPositionObject(pointer);
     let edge = this.body.edges[this.edgeBeingEditedId];
-
     // if the node that was dragged is not a control node, return
     if (this.selectedControlNode === undefined) {
       return;
     }
 
+    // we use the selection to find the node that is being dragged. We explicitly DEselect the control node here.
+    this.selectionHandler.unselectAll();
     let overlappingNodeIds = this.selectionHandler._getAllNodesOverlappingWith(pointerObj);
     let node = undefined;
     for (let i = overlappingNodeIds.length-1; i >= 0; i--) {
@@ -898,7 +908,6 @@ class ManipulationSystem {
         break;
       }
     }
-
     // perform the connection
     if (node !== undefined && this.selectedControlNode !== undefined) {
       if (node.isCluster === true) {
