@@ -49,7 +49,7 @@ class NodesHandler {
         background: 'none',
         strokeWidth: 0, // px
         strokeColor: '#ffffff',
-        align: 'horizontal'
+        align: 'center'
       },
       group: undefined,
       hidden: false,
@@ -96,6 +96,7 @@ class NodesHandler {
       shapeProperties: {
         borderDashes: false, // only for borders
         borderRadius: 6,     // only for box shape
+        interpolation: true,  // only for image and circularImage shapes
         useImageSize: false,  // only for image and circularImage shapes
         useBorderWithImage: false  // only for image shape
       },
