@@ -49,7 +49,7 @@ class Edge {
 
     this.connected = false;
 
-    this.labelModule = new Label(this.body, this.options);
+    this.labelModule = new Label(this.body, this.options, true /* It's an edge label */);
 
     this.setOptions(options);
   }
@@ -142,9 +142,9 @@ class Edge {
     if (newOptions.arrows !== undefined && newOptions.arrows !== null) {
       if (typeof newOptions.arrows === 'string') {
         let arrows = newOptions.arrows.toLowerCase();
-        if (arrows.indexOf("to")     != -1) {parentOptions.arrows.to.enabled     = true;}
-        if (arrows.indexOf("middle") != -1) {parentOptions.arrows.middle.enabled = true;}
-        if (arrows.indexOf("from")   != -1) {parentOptions.arrows.from.enabled   = true;}
+        parentOptions.arrows.to.enabled     = arrows.indexOf("to")     != -1;
+        parentOptions.arrows.middle.enabled = arrows.indexOf("middle") != -1;
+        parentOptions.arrows.from.enabled   = arrows.indexOf("from")   != -1;
       }
       else if (typeof newOptions.arrows === 'object') {
         util.mergeOptions(parentOptions.arrows, newOptions.arrows, 'to',     allowDeletion, globalOptions.arrows);
@@ -379,8 +379,8 @@ class Edge {
     let arrowData = {};
 
     // restore edge targets to defaults
-    this.edgeType.fromPoint = this.from;
-    this.edgeType.toPoint = this.to;
+    this.edgeType.fromPoint = this.edgeType.from;
+    this.edgeType.toPoint = this.edgeType.to;
 
     // from and to arrows give a different end point for edges. we set them here
     if (this.options.arrows.from.enabled === true) {
