@@ -67,7 +67,7 @@ class Node {
     this.selected = false;
     this.hover = false;
 
-    this.labelModule = new Label(this.body, this.options);
+    this.labelModule = new Label(this.body, this.options, false /* Not edge label */);
     this.setOptions(options);
   }
 
