@@ -103,8 +103,37 @@ class CircleImageBase extends NodeBase {
       // draw shadow if enabled
       this.enableShadow(ctx);
 
-      // draw image
-      ctx.drawImage(this.imageObj, this.left, this.top, this.width, this.height);
+      let factor = (this.imageObj.width / this.width) / this.body.view.scale;
+      if (factor > 2 && this.options.shapeProperties.interpolation === true) {
+        let w = this.imageObj.width;
+        let h = this.imageObj.height;
+        var can2 = document.createElement('canvas');
+        can2.width = w;
+        can2.height = w;
+        var ctx2 = can2.getContext('2d');
+
+        factor *= 0.5;
+        w *= 0.5;
+        h *= 0.5;
+        ctx2.drawImage(this.imageObj, 0, 0, w, h);
+
+        let distance = 0;
+        let iterations = 1;
+        while (factor > 2 && iterations < 4) {
+          ctx2.drawImage(can2, distance, 0, w, h, distance+w, 0, w/2, h/2);
+          distance += w;
+          factor *= 0.5;
+          w *= 0.5;
+          h *= 0.5;
+          iterations += 1;
+        }
+        ctx.drawImage(can2, distance, 0, w, h, this.left, this.top, this.width, this.height);
+      }
+      else {
+        // draw image
+        ctx.drawImage(this.imageObj, this.left, this.top, this.width, this.height);
+      }
+
 
       // disable shadows for other elements.
       this.disableShadow(ctx);
