@@ -118,8 +118,9 @@ Core.prototype._create = function (container) {
   // create event listeners for all interesting events, these events will be
   // emitted via emitter
   this.hammer = new Hammer(this.dom.root);
-  this.hammer.get('pinch').set({enable: true});
-  this.hammer.get('pan').set({threshold:5, direction:30}); // 30 is ALL_DIRECTIONS in hammer.
+  var pinchRecognizer = this.hammer.get('pinch').set({enable: true});
+  hammerUtil.disablePreventDefaultVertically(pinchRecognizer);
+  this.hammer.get('pan').set({threshold:5, direction: Hammer.DIRECTION_HORIZONTAL});
   this.listeners = {};
 
   var events = [
@@ -221,10 +222,18 @@ Core.prototype.setOptions = function (options) {
     var fields = [
       'width', 'height', 'minHeight', 'maxHeight', 'autoResize',
       'start', 'end', 'clickToUse', 'dataAttributes', 'hiddenDates',
-      'locale', 'locales', 'moment',
+      'locale', 'locales', 'moment', 'rtl',
       'throttleRedraw'
     ];
     util.selectiveExtend(fields, this.options, options);
+
+
+    if (this.options.rtl) {
+      var contentContainer = this.dom.leftContainer;
+      this.dom.leftContainer = this.dom.rightContainer;
+      this.dom.rightContainer = contentContainer;
+      this.dom.container.style.direction = "rtl";
+      this.dom.backgroundVertical.className = 'vis-panel vis-background vis-vertical-rtl';    }  
 
     this.options.orientation = {item:undefined,axis:undefined};
     if ('orientation' in options) {
@@ -528,7 +537,6 @@ Core.prototype.fit = function(options) {
   var interval = range.max - range.min;
   var min = new Date(range.min.valueOf() - interval * 0.01);
   var max = new Date(range.max.valueOf() + interval * 0.01);
-
   var animation = (options && options.animation !== undefined) ? options.animation : true;
   this.range.setRange(min, max, animation);
 };
@@ -632,7 +640,7 @@ Core.prototype._redraw = function() {
   var props = this.props;
   var dom = this.dom;
 
-  if (!dom|| !dom.container || dom.container.clientWidth == 0 ) return;// when destroyed, or invisible
+  if (!dom || !dom.container || dom.root.offsetWidth == 0) return; // when destroyed, or invisible
 
   DateUtil.updateHiddenDates(this.options.moment, this.body, this.options.hiddenDates);
 
@@ -767,6 +775,12 @@ Core.prototype._redraw = function() {
   dom.shadowBottomLeft.style.visibility   = visibilityBottom;
   dom.shadowTopRight.style.visibility     = visibilityTop;
   dom.shadowBottomRight.style.visibility  = visibilityBottom;
+
+  // enable/disable vertical panning
+  var contentsOverflow = this.props.center.height > this.props.centerContainer.height;
+  this.hammer.get('pan').set({
+    direction: contentsOverflow ? Hammer.DIRECTION_ALL : Hammer.DIRECTION_HORIZONTAL
+  });
 
   // redraw all components
   this.components.forEach(function (component) {
