@@ -11,11 +11,12 @@ var CurrentTime = require('./component/CurrentTime');
 var CustomTime = require('./component/CustomTime');
 var LineGraph = require('./component/LineGraph');
 
-var Configurator = require('../shared/Configurator');
-var Validator = require('../shared/Validator').default;
 var printStyle = require('../shared/Validator').printStyle;
 var allOptions = require('./optionsGraph2d').allOptions;
 var configureOptions = require('./optionsGraph2d').configureOptions;
+
+import Configurator from '../shared/Configurator';
+import Validator from '../shared/Validator';
 
 /**
  * Create a timeline visualization
