@@ -25,6 +25,7 @@ function Range(body, options) {
 
   // default options
   this.defaultOptions = {
+    rtl: false,
     start: null,
     end: null,
     moment: moment,
@@ -37,7 +38,6 @@ function Range(body, options) {
     zoomMax: 1000 * 60 * 60 * 24 * 365 * 10000  // milliseconds
   };
   this.options = util.extend({}, this.defaultOptions);
-
   this.props = {
     touch: {}
   };
@@ -81,7 +81,7 @@ Range.prototype.setOptions = function (options) {
     // copy the options that we know
     var fields = [
       'direction', 'min', 'max', 'zoomMin', 'zoomMax', 'moveable', 'zoomable',
-      'moment', 'activate', 'hiddenDates', 'zoomKey'
+      'moment', 'activate', 'hiddenDates', 'zoomKey', 'rtl'
     ];
     util.selectiveExtend(fields, this.options, options);
 
@@ -411,7 +411,13 @@ Range.prototype._onDrag = function (event) {
   interval -= duration;
 
   var width = (direction == 'horizontal') ? this.body.domProps.center.width : this.body.domProps.center.height;
-  var diffRange = -delta / width * interval;
+
+  if (this.options.rtl) {
+    var diffRange = delta / width * interval;
+  } else {
+     var diffRange = -delta / width * interval;
+  }
+
   var newStart = this.props.touch.start + diffRange;
   var newEnd = this.props.touch.end + diffRange;
 
@@ -513,7 +519,7 @@ Range.prototype._onMouseWheel = function(event) {
     }
 
     // calculate center, the date to zoom around
-    var pointer = getPointer({x: event.clientX, y: event.clientY}, this.body.dom.center);
+    var pointer = this.getPointer({x: event.clientX, y: event.clientY}, this.body.dom.center);
     var pointerDate = this._pointerToDate(pointer);
 
     this.zoom(scale, pointerDate, delta);
@@ -549,7 +555,7 @@ Range.prototype._onPinch = function (event) {
   this.props.touch.allowDragging = false;
 
   if (!this.props.touch.center) {
-    this.props.touch.center = getPointer(event.center, this.body.dom.center);
+    this.props.touch.center = this.getPointer(event.center, this.body.dom.center);
   }
 
   var scale = 1 / (event.scale + this.scaleOffset);
@@ -594,7 +600,11 @@ Range.prototype._isInsideRange = function(event) {
   // calculate the time where the mouse is, check whether inside
   // and no scroll action should happen.
   var clientX = event.center ? event.center.x : event.clientX;
-  var x = clientX - util.getAbsoluteLeft(this.body.dom.centerContainer);
+  if (this.options.rtl) {
+    var x = clientX - util.getAbsoluteLeft(this.body.dom.centerContainer);
+  } else {
+    var x = util.getAbsoluteRight(this.body.dom.centerContainer) - clientX;
+  }
   var time = this.body.util.toTime(x);
 
   return time >= this.start && time <= this.end;
@@ -629,11 +639,18 @@ Range.prototype._pointerToDate = function (pointer) {
  * @return {{x: Number, y: Number}} pointer
  * @private
  */
-function getPointer (touch, element) {
-  return {
-    x: touch.x - util.getAbsoluteLeft(element),
-    y: touch.y - util.getAbsoluteTop(element)
-  };
+Range.prototype.getPointer = function (touch, element) {
+  if (this.options.rtl) {
+    return {
+      x: util.getAbsoluteRight(element) - touch.x,
+      y: touch.y - util.getAbsoluteTop(element)
+    };
+  } else {
+    return {
+      x: touch.x - util.getAbsoluteLeft(element),
+      y: touch.y - util.getAbsoluteTop(element)
+    };
+  }
 }
 
 /**
