@@ -37,8 +37,7 @@ exports.orderByEnd = function(items) {
  *            items having a top===null will be re-stacked
  */
 exports.stack = function(items, margin, force) {
-  var i, iMax;
-
+  var i, iMax;  
   if (force) {
     // reset top position of all items
     for (i = 0, iMax = items.length; i < iMax; i++) {
@@ -59,7 +58,7 @@ exports.stack = function(items, margin, force) {
         var collidingItem = null;
         for (var j = 0, jj = items.length; j < jj; j++) {
           var other = items[j];
-          if (other.top !== null && other !== item && other.stack && exports.collision(item, other, margin.item)) {
+          if (other.top !== null && other !== item && other.stack && exports.collision(item, other, margin.item, other.options.rtl)) {
             collidingItem = other;
             break;
           }
@@ -112,11 +111,19 @@ exports.nostack = function(items, margin, subgroups) {
  * @param {{horizontal: number, vertical: number}} margin
  *                          An object containing a horizontal and vertical
  *                          minimum required margin.
+ * @param {boolean} rtl
  * @return {boolean}        true if a and b collide, else false
  */
-exports.collision = function(a, b, margin) {
-  return ((a.left - margin.horizontal + EPSILON)       < (b.left + b.width) &&
-      (a.left + a.width + margin.horizontal - EPSILON) > b.left &&
-      (a.top - margin.vertical + EPSILON)              < (b.top + b.height) &&
-      (a.top + a.height + margin.vertical - EPSILON)   > b.top);
+exports.collision = function(a, b, margin, rtl) {
+  if (rtl) {
+    return  ((a.right - margin.horizontal + EPSILON)  < (b.right + b.width) &&
+    (a.right + a.width + margin.horizontal - EPSILON) > b.right &&
+    (a.top - margin.vertical + EPSILON)              < (b.top + b.height) &&
+    (a.top + a.height + margin.vertical - EPSILON)   > b.top);
+  } else {
+    return ((a.left - margin.horizontal + EPSILON)   < (b.left + b.width) &&
+    (a.left + a.width + margin.horizontal - EPSILON) > b.left &&
+    (a.top - margin.vertical + EPSILON)              < (b.top + b.height) &&
+    (a.top + a.height + margin.vertical - EPSILON)   > b.top);
+  }
 };
