@@ -11,11 +11,13 @@ var CurrentTime = require('./component/CurrentTime');
 var CustomTime = require('./component/CustomTime');
 var ItemSet = require('./component/ItemSet');
 
-var Configurator = require('../shared/Configurator');
-var Validator = require('../shared/Validator').default;
 var printStyle = require('../shared/Validator').printStyle;
 var allOptions = require('./optionsTimeline').allOptions;
 var configureOptions = require('./optionsTimeline').configureOptions;
+
+import Configurator from '../shared/Configurator';
+import Validator from '../shared/Validator';
+
 
 /**
  * Create a timeline visualization
@@ -27,6 +29,7 @@ var configureOptions = require('./optionsTimeline').configureOptions;
  * @extends Core
  */
 function Timeline (container, items, groups, options) {
+
   if (!(this instanceof Timeline)) {
     throw new SyntaxError('Constructor must be called with the new operator');
   }
@@ -50,7 +53,7 @@ function Timeline (container, items, groups, options) {
       axis: 'bottom',   // axis orientation: 'bottom', 'top', or 'both'
       item: 'bottom'    // not relevant
     },
-
+    rtl: false,
     moment: moment,
 
     width: null,
@@ -105,7 +108,7 @@ function Timeline (container, items, groups, options) {
   this.components.push(this.currentTime);
 
   // item set
-  this.itemSet = new ItemSet(this.body);
+  this.itemSet = new ItemSet(this.body, this.options);
   this.components.push(this.itemSet);
 
   this.itemsData = null;      // DataSet
@@ -133,7 +136,7 @@ function Timeline (container, items, groups, options) {
         }
 
         var start = me.options.start != undefined ? me.options.start : range.min;
-        var end = me.options.end != undefined ? me.options.end : range.max;
+        var end   = me.options.end   != undefined ? me.options.end   : range.max;
 
         me.setWindow(start, end, {animation: false});
       }
@@ -189,6 +192,7 @@ Timeline.prototype.redraw = function() {
 Timeline.prototype.setOptions = function (options) {
   // validate options
   let errorFound = Validator.validate(options, allOptions);
+
   if (errorFound === true) {
     console.log('%cErrors have been found in the supplied options object.', printStyle);
   }
@@ -371,8 +375,19 @@ Timeline.prototype.focus = function(id, options) {
  */
 Timeline.prototype.fit = function (options) {
   var animation = (options && options.animation !== undefined) ? options.animation : true;
-  var range = this.getItemRange();
-  this.range.setRange(range.min, range.max, animation);
+  var range;
+
+  var dataset = this.itemsData && this.itemsData.getDataSet();
+  if (dataset.length === 1 && dataset.get()[0].end === undefined) {
+    // a single item -> don't fit, just show a range around the item from -4 to +3 days
+    range = this.getDataRange();
+    this.moveTo(range.min.valueOf(), {animation});
+  }
+  else {
+    // exactly fit the items (plus a small margin)
+    range = this.getItemRange();
+    this.range.setRange(range.min, range.max, animation);
+  }
 };
 
 /**
@@ -411,16 +426,22 @@ Timeline.prototype.getItemRange = function () {
 
       var start = getStart(item);
       var end = getEnd(item);
+      
+      if (this.options.rtl) {
+        var startSide  = start - (item.getWidthRight()  + 10) * factor;
+        var endSide = end   + (item.getWidthLeft() + 10) * factor;
+      } else {
+        var startSide  = start - (item.getWidthLeft()  + 10) * factor;
+        var endSide = end   + (item.getWidthRight() + 10) * factor;
+      }
+        
 
-      var left  = start - (item.getWidthLeft()  + 10) * factor;
-      var right = end   + (item.getWidthRight() + 10) * factor;
-
-      if (left < min) {
-        min = left;
+      if (startSide < min) {
+        min = startSide;
         minItem = item;
       }
-      if (right > max) {
-        max = right;
+      if (endSide > max) {
+        max = endSide;
         maxItem = item;
       }
     }.bind(this));
@@ -431,8 +452,13 @@ Timeline.prototype.getItemRange = function () {
       var delta = this.props.center.width - lhs - rhs;  // px
 
       if (delta > 0) {
-        min = getStart(minItem) - lhs * interval / delta; // ms
-        max = getEnd(maxItem)   + rhs * interval / delta; // ms
+        if (this.options.rtl) {
+          min = getStart(minItem) - rhs * interval / delta; // ms
+          max = getEnd(maxItem)   + lhs * interval / delta; // ms
+        } else {
+          min = getStart(minItem) - lhs * interval / delta; // ms
+          max = getEnd(maxItem)   + rhs * interval / delta; // ms
+        }
       }
     }
   }
@@ -480,7 +506,11 @@ Timeline.prototype.getDataRange = function() {
 Timeline.prototype.getEventProperties = function (event) {
   var clientX = event.center ? event.center.x : event.clientX;
   var clientY = event.center ? event.center.y : event.clientY;
-  var x = clientX - util.getAbsoluteLeft(this.dom.centerContainer);
+  if (this.options.rtl) {
+    var x = util.getAbsoluteRight(this.dom.centerContainer) - clientX;
+  } else {
+    var x = clientX - util.getAbsoluteLeft(this.dom.centerContainer);
+  }
   var y = clientY - util.getAbsoluteTop(this.dom.centerContainer);
 
   var item  = this.itemSet.itemFromTarget(event);
