@@ -16,6 +16,7 @@ function CurrentTime (body, options) {
 
   // default options
   this.defaultOptions = {
+    rtl: false,
     showCurrentTime: true,
 
     moment: moment,
@@ -64,7 +65,7 @@ CurrentTime.prototype.destroy = function () {
 CurrentTime.prototype.setOptions = function(options) {
   if (options) {
     // copy all options that we know
-    util.selectiveExtend(['showCurrentTime', 'moment', 'locale', 'locales'], this.options, options);
+    util.selectiveExtend(['rtl', 'showCurrentTime', 'moment', 'locale', 'locales'], this.options, options);
   }
 };
 
@@ -99,7 +100,11 @@ CurrentTime.prototype.redraw = function() {
     var title = locale.current + ' ' + locale.time + ': ' + now.format('dddd, MMMM Do YYYY, H:mm:ss');
     title = title.charAt(0).toUpperCase() + title.substring(1);
 
-    this.bar.style.left = x + 'px';
+    if (this.options.rtl) {
+      this.bar.style.right = x + 'px';
+    } else {
+      this.bar.style.left = x + 'px';
+    }
     this.bar.title = title;
   }
   else {
@@ -129,6 +134,7 @@ CurrentTime.prototype.start = function() {
     if (interval > 1000) interval = 1000;
 
     me.redraw();
+    me.body.emitter.emit('currentTimeTick');
 
     // start a renderTimer to adjust for the new time
     me.currentTimeTimer = setTimeout(update, interval);
