@@ -29,12 +29,12 @@ function DataAxis (body, options, svg, linegraphOptions) {
     alignZeros: true,
     left:{
       range: {min:undefined,max:undefined},
-      format: function (value) {return ''+Number.parseFloat(value.toPrecision(3));},
+      format: function (value) {return ''+parseFloat(value.toPrecision(3));},
       title: {text:undefined,style:undefined}
     },
     right:{
       range: {min:undefined,max:undefined},
-      format: function (value) {return ''+Number.parseFloat(value.toPrecision(3));},
+      format: function (value) {return ''+parseFloat(value.toPrecision(3));},
       title: {text:undefined,style:undefined}
     }
   };
@@ -212,11 +212,11 @@ DataAxis.prototype._cleanupIcons = function() {
 DataAxis.prototype.show = function() {
   this.hidden = false;
   if (!this.dom.frame.parentNode) {
-    if (this.options.orientation === 'left') {
+    if (this.options.rtl) {
       this.body.dom.left.appendChild(this.dom.frame);
     }
     else {
-      this.body.dom.right.appendChild(this.dom.frame);
+      this.body.dom.left.appendChild(this.dom.frame);
     }
   }
 
