@@ -209,8 +209,8 @@ Group.prototype.redraw = function(range, margin, restack) {
   }
   else {
     // no custom order function, lazy stacking
-    this.visibleItems = this._updateVisibleItems(this.orderedItems, this.visibleItems, range);
 
+    this.visibleItems = this._updateVisibleItems(this.orderedItems, this.visibleItems, range);
     if (this.itemSet.options.stack) { // TODO: ugly way to access options...
       stack.stack(this.visibleItems, margin, restack);
     }
@@ -225,10 +225,9 @@ Group.prototype.redraw = function(range, margin, restack) {
   // calculate actual size and position
   var foreground = this.dom.foreground;
   this.top = foreground.offsetTop;
-  this.left = foreground.offsetLeft;
+  this.right = foreground.offsetLeft;
   this.width = foreground.offsetWidth;
   resized = util.updateProperty(this, 'height', height) || resized;
-
   // recalculate size of label
   resized = util.updateProperty(this.props.label, 'width', this.dom.inner.clientWidth) || resized;
   resized = util.updateProperty(this.props.label, 'height', this.dom.inner.clientHeight) || resized;
