@@ -73,7 +73,8 @@ TimeAxis.prototype.setOptions = function(options) {
       'maxMinorChars',
       'hiddenDates',
       'timeAxis',
-      'moment'
+      'moment',
+      'rtl'
     ], this.options, options);
 
     // deep copy the format options
@@ -183,7 +184,6 @@ TimeAxis.prototype.redraw = function () {
   else {
     this.body.dom.backgroundVertical.appendChild(background)
   }
-
   return this._isResized() || parentChanged;
 };
 
@@ -336,7 +336,13 @@ TimeAxis.prototype._repaintMinorText = function (x, text, orientation, className
   label.childNodes[0].nodeValue = text;
 
   label.style.top = (orientation == 'top') ? (this.props.majorLabelHeight + 'px') : '0';
-  label.style.left = x + 'px';
+
+  if (this.options.rtl) {
+    label.style.left = "";
+    label.style.right = x + 'px';
+  } else {
+    label.style.left = x + 'px';
+  };
   label.className = 'vis-text vis-minor ' + className;
   //label.title = title;  // TODO: this is a heavy operation
 
@@ -370,7 +376,12 @@ TimeAxis.prototype._repaintMajorText = function (x, text, orientation, className
   //label.title = title; // TODO: this is a heavy operation
 
   label.style.top = (orientation == 'top') ? '0' : (this.props.minorLabelHeight  + 'px');
-  label.style.left = x + 'px';
+  if (this.options.rtl) {
+    label.style.left = "";
+    label.style.right = x + 'px';
+  } else {
+    label.style.left = x + 'px';
+  };
 
   return label;
 };
@@ -402,10 +413,17 @@ TimeAxis.prototype._repaintMinorLine = function (x, width, orientation, classNam
     line.style.top = this.body.domProps.top.height + 'px';
   }
   line.style.height = props.minorLineHeight + 'px';
-  line.style.left = (x - props.minorLineWidth / 2) + 'px';
+  if (this.options.rtl) {
+    line.style.left = "";
+    line.style.right = (x - props.minorLineWidth / 2) + 'px';
+    line.className = 'vis-grid vis-vertical-rtl vis-minor ' + className;
+  } else {
+    line.style.left = (x - props.minorLineWidth / 2) + 'px';
+    line.className = 'vis-grid vis-vertical vis-minor ' + className;
+  };
   line.style.width = width + 'px';
 
-  line.className = 'vis-grid vis-vertical vis-minor ' + className;
+  
 
   return line;
 };
@@ -436,11 +454,18 @@ TimeAxis.prototype._repaintMajorLine = function (x, width, orientation, classNam
   else {
     line.style.top = this.body.domProps.top.height + 'px';
   }
-  line.style.left = (x - props.majorLineWidth / 2) + 'px';
+
+  if (this.options.rtl) {
+    line.style.left = "";
+    line.style.right = (x - props.majorLineWidth / 2) + 'px';
+    line.className = 'vis-grid vis-vertical-rtl vis-major ' + className;
+  } else {
+    line.style.left = (x - props.majorLineWidth / 2) + 'px';
+    line.className = 'vis-grid vis-vertical vis-major ' + className;
+  }
+
   line.style.height = props.majorLineHeight + 'px';
   line.style.width = width  + 'px';
-
-  line.className = 'vis-grid vis-vertical vis-major ' + className;
 
   return line;
 };
