@@ -22,7 +22,7 @@ function BoxItem (data, conversion, options) {
       height: 0
     }
   };
-
+  this.options = options;
   // validate data
   if (data) {
     if (data.start == undefined) {
@@ -168,28 +168,60 @@ BoxItem.prototype.hide = function() {
 BoxItem.prototype.repositionX = function() {
   var start = this.conversion.toScreen(this.data.start);
   var align = this.options.align;
-  var left;
 
   // calculate left position of the box
   if (align == 'right') {
-    this.left = start - this.width;
+    if (this.options.rtl) {
+      this.right = start - this.width;
+
+      // reposition box, line, and dot
+      this.dom.box.style.right = this.right + 'px';
+      this.dom.line.style.right = (start - this.props.line.width) + 'px';
+      this.dom.dot.style.right = (start - this.props.line.width / 2 - this.props.dot.width / 2) + 'px';
+    } else {
+      this.left = start - this.width;
+
+      // reposition box, line, and dot
+      this.dom.box.style.left = this.left + 'px';
+      this.dom.line.style.left = (start - this.props.line.width) + 'px';
+      this.dom.dot.style.left = (start - this.props.line.width / 2 - this.props.dot.width / 2) + 'px';
+    }
   }
   else if (align == 'left') {
-    this.left = start;
+    if (this.options.rtl) {
+      this.right = start;
+
+      // reposition box, line, and dot
+      this.dom.box.style.right = this.right + 'px';
+      this.dom.line.style.right = start + 'px';
+      this.dom.dot.style.right = (start + this.props.line.width / 2 - this.props.dot.width / 2) + 'px';
+    } else {
+      this.left = start;
+
+      // reposition box, line, and dot
+      this.dom.box.style.left = this.left + 'px';
+      this.dom.line.style.left = start + 'px';
+      this.dom.dot.style.left = (start + this.props.line.width / 2 - this.props.dot.width / 2) + 'px';
+    }
   }
   else {
     // default or 'center'
-    this.left = start - this.width / 2;
+    if (this.options.rtl) {
+      this.right = start - this.width / 2;
+
+      // reposition box, line, and dot
+      this.dom.box.style.right = this.right + 'px';
+      this.dom.line.style.right = (start - this.props.line.width) + 'px';
+      this.dom.dot.style.right = (start - this.props.dot.width / 2) + 'px';
+    } else {
+      this.left = start - this.width / 2;
+
+      // reposition box, line, and dot
+      this.dom.box.style.left = this.left + 'px';
+      this.dom.line.style.left = (start - this.props.line.width / 2) + 'px';
+      this.dom.dot.style.left = (start - this.props.dot.width / 2) + 'px';
+    }
   }
-
-  // reposition box
-  this.dom.box.style.left = this.left + 'px';
-
-  // reposition line
-  this.dom.line.style.left = (start - this.props.line.width / 2) + 'px';
-
-  // reposition dot
-  this.dom.dot.style.left = (start - this.props.dot.width / 2) + 'px';
 };
 
 /**
