@@ -23,6 +23,7 @@ function Item (data, conversion, options) {
   this.dirty = true;
 
   this.top = null;
+  this.right = null;
   this.left = null;
   this.width = null;
   this.height = null;
@@ -154,7 +155,12 @@ Item.prototype._repaintDeleteButton = function (anchor) {
     var me = this;
 
     var deleteButton = document.createElement('div');
-    deleteButton.className = 'vis-delete';
+
+    if (this.options.rtl) {
+      deleteButton.className = 'vis-delete-rtl';
+    } else {
+      deleteButton.className = 'vis-delete';
+    }
     deleteButton.title = 'Delete this item';
 
     // TODO: be able to destroy the delete button
