@@ -19,10 +19,11 @@ function PointItem (data, conversion, options) {
     },
     content: {
       height: 0,
-      marginLeft: 0
+      marginLeft: 0,
+      marginRight: 0
     }
   };
-
+  this.options = options;
   // validate data
   if (data) {
     if (data.start == undefined) {
@@ -117,7 +118,11 @@ PointItem.prototype.redraw = function() {
     this.props.content.height = dom.content.offsetHeight;
 
     // resize contents
-    dom.content.style.marginLeft = 2 * this.props.dot.width + 'px';
+    if (this.options.rtl) {
+      dom.content.style.marginRight = 2 * this.props.dot.width + 'px';
+    } else {
+      dom.content.style.marginLeft = 2 * this.props.dot.width + 'px';
+    }
     //dom.content.style.marginRight = ... + 'px'; // TODO: margin right
 
     // recalculate size
@@ -126,7 +131,11 @@ PointItem.prototype.redraw = function() {
 
     // reposition the dot
     dom.dot.style.top = ((this.height - this.props.dot.height) / 2) + 'px';
-    dom.dot.style.left = (this.props.dot.width / 2) + 'px';
+    if (this.options.rtl) {
+      dom.dot.style.right = (this.props.dot.width / 2) + 'px';
+    } else {
+      dom.dot.style.left = (this.props.dot.width / 2) + 'px';
+    }
 
     this.dirty = false;
   }
@@ -164,10 +173,17 @@ PointItem.prototype.hide = function() {
 PointItem.prototype.repositionX = function() {
   var start = this.conversion.toScreen(this.data.start);
 
-  this.left = start - this.props.dot.width;
+  if (this.options.rtl) {
+    this.right = start - this.props.dot.width;
 
-  // reposition point
-  this.dom.point.style.left = this.left + 'px';
+    // reposition point
+    this.dom.point.style.right = this.right + 'px';
+  } else {
+    this.left = start - this.props.dot.width;
+
+    // reposition point
+    this.dom.point.style.left = this.left + 'px';
+  }
 };
 
 /**
@@ -177,7 +193,6 @@ PointItem.prototype.repositionX = function() {
 PointItem.prototype.repositionY = function() {
   var orientation = this.options.orientation.item;
   var point = this.dom.point;
-
   if (orientation == 'top') {
     point.style.top = this.top + 'px';
   }
@@ -199,7 +214,7 @@ PointItem.prototype.getWidthLeft = function () {
  * @return {number}
  */
 PointItem.prototype.getWidthRight = function () {
-  return this.width - this.props.dot.width;
+  return this.props.dot.width;
 };
 
 module.exports = PointItem;
