@@ -18,7 +18,7 @@ function RangeItem (data, conversion, options) {
     }
   };
   this.overflow = false; // if contents can overflow (css styling), this flag is set to true
-
+  this.options = options;
   // validate data
   if (data) {
     if (data.start == undefined) {
@@ -123,7 +123,6 @@ RangeItem.prototype.redraw = function() {
 
     this.dirty = false;
   }
-
   this._repaintDeleteButton(dom.box);
   this._repaintDragLeft();
   this._repaintDragRight();
@@ -168,7 +167,7 @@ RangeItem.prototype.repositionX = function(limitSize) {
   var parentWidth = this.parent.width;
   var start = this.conversion.toScreen(this.data.start);
   var end = this.conversion.toScreen(this.data.end);
-  var contentLeft;
+  var contentStartPosition;
   var contentWidth;
 
   // limit the width of the range, as browsers cannot draw very wide divs
@@ -183,7 +182,11 @@ RangeItem.prototype.repositionX = function(limitSize) {
   var boxWidth = Math.max(end - start, 1);
 
   if (this.overflow) {
-    this.left = start;
+    if (this.options.rtl) {
+      this.right = start;
+    } else {
+      this.left = start;
+    }
     this.width = boxWidth + this.props.content.width;
     contentWidth = this.props.content.width;
 
@@ -192,46 +195,71 @@ RangeItem.prototype.repositionX = function(limitSize) {
     //       So no re-stacking needed, which is nicer for the eye;
   }
   else {
-    this.left = start;
+    if (this.options.rtl) {
+      this.right = start;
+    } else {
+      this.left = start;
+    }
     this.width = boxWidth;
     contentWidth = Math.min(end - start, this.props.content.width);
   }
 
-  this.dom.box.style.left = this.left + 'px';
+  if (this.options.rtl) {
+    this.dom.box.style.right = this.right + 'px';
+  } else {
+    this.dom.box.style.left = this.left + 'px'; 
+  }
   this.dom.box.style.width = boxWidth + 'px';
 
   switch (this.options.align) {
     case 'left':
-      this.dom.content.style.left = '0';
+      if (this.options.rtl) {
+        this.dom.content.style.right = '0';
+      } else {
+        this.dom.content.style.left = '0';
+      }
       break;
 
     case 'right':
-      this.dom.content.style.left = Math.max((boxWidth - contentWidth), 0) + 'px';
+      if (this.options.rtl) {
+        this.dom.content.style.right = Math.max((boxWidth - contentWidth), 0) + 'px';
+      } else {
+        this.dom.content.style.left = Math.max((boxWidth - contentWidth), 0) + 'px';
+      }
       break;
 
     case 'center':
-      this.dom.content.style.left = Math.max((boxWidth - contentWidth) / 2, 0) + 'px';
+      if (this.options.rtl) {
+        this.dom.content.style.right = Math.max((boxWidth - contentWidth) / 2, 0) + 'px';
+      } else {
+        this.dom.content.style.left = Math.max((boxWidth - contentWidth) / 2, 0) + 'px';
+      }
+      
       break;
 
     default: // 'auto'
       // when range exceeds left of the window, position the contents at the left of the visible area
       if (this.overflow) {
         if (end > 0) {
-          contentLeft = Math.max(-start, 0);
+          contentStartPosition = Math.max(-start, 0);
         }
         else {
-          contentLeft = -contentWidth; // ensure it's not visible anymore
+          contentStartPosition = -contentWidth; // ensure it's not visible anymore
         }
       }
       else {
         if (start < 0) {
-          contentLeft = -start;
+          contentStartPosition = -start;
         }
         else {
-          contentLeft = 0;
+          contentStartPosition = 0;
         }
       }
-      this.dom.content.style.left = contentLeft + 'px';
+      if (this.options.rtl) {
+        this.dom.content.style.right = contentStartPosition + 'px';
+      } else {
+        this.dom.content.style.left = contentStartPosition + 'px';
+      }
   }
 };
 
