@@ -15,7 +15,6 @@ let dom = 'dom';
 let moment = 'moment';
 let any = 'any';
 
-
 let allOptions = {
   configure: {
     enabled: {boolean},
@@ -26,6 +25,7 @@ let allOptions = {
 
   //globals :
   align: {string},
+  rtl: {boolean, 'undefined': 'undefined'},
   autoResize: {boolean},
   throttleRedraw: {number},
   clickToUse: {boolean},
@@ -143,6 +143,7 @@ let allOptions = {
 let configureOptions = {
   global: {
     align:  ['center', 'left', 'right'],
+    direction:  false,
     autoResize: true,
     throttleRedraw: [10, 0, 1000, 10],
     clickToUse: false,
